+++ CHANGELOG.md
@@ -11,11 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New hook makes View properties directly accessible in TemplateFiles (e.g. when rendering field templates)
 
 ### Changed
-- Various minor optimizations, some code cleanup, and a few improvements to comments.
-
-### Changed
 - Wireframe::getController() is now a public method that can return current Controller instance, the Controller instance for provided page, or a Controller instance for provided Page and template name.
 - Visibility of following methods was changed from public to protected: Wireframe::___checkRedirects(), Wireframe::___redirect(), Wireframe::___initView(), Wireframe::___initController(), \Wireframe\Config::getCreateDirectoriesField().
+- Wireframe no longer caches partials unnecessarily, plus new Partial objects are automatically wired.
+- Various minor optimizations, some code cleanup, and a few improvements to comments.
 
 ## [0.12.0] - 2020-08-24